fix: int type 에러

+++ src/main/java/lotto/model/WinningMoney.java
@@ -11,13 +11,13 @@ public enum WinningMoney {
 	FIFTH(3, 5_000), 
 	FOURTH(4, 50_000), 
 	THIRD(5, 1_500_000), 
-	SECOND(5, 30_000_000), 
+	SECOND(5, 30_000_000),
 	FIRST(6, 2_000_000_000);
 
-	private static final Map<Integer, WinningMoney> WINNING_MONEY_MAP = Collections
+	private static final Map<Long, WinningMoney> WINNING_MONEY_MAP = Collections
 			.unmodifiableMap(Stream.of(values()).filter(WinningMoney::isNotSecondPlace)
 					.collect(Collectors.toMap(WinningMoney::getMatchCount, Function.identity())));
-	private static final int DEFAULT_WINNING_MONEY_KEY = 0;
+	private static final long DEFAULT_WINNING_MONEY_KEY = 0L;
 
 	private final int matchCount;
 	private final int winningMoney;
@@ -27,25 +27,25 @@ public enum WinningMoney {
 		this.winningMoney = winningMoney;
 	}
 
-	public int getMatchCount() {
+	public long getMatchCount() {
 		return matchCount;
 	}
 
-	public int getWinningMoney() {
+	public long getWinningMoney() {
 		return winningMoney;
 	}
 
-	public int winningMoney(int count) {
-		return winningMoney * count;
+	public long winningMoney(int count) {
+		return (long)winningMoney * count;
 	}
 
-	public static WinningMoney find(int matchCount) {
+	public static WinningMoney find(long matchCount) {
 		if (WINNING_MONEY_MAP.containsKey(matchCount)) {
 			return WINNING_MONEY_MAP.get(matchCount);
 		}
 		return WINNING_MONEY_MAP.get(DEFAULT_WINNING_MONEY_KEY);
 	}
-	
+
 	public boolean isSecondPlace() {
 		return this == SECOND;
 	}
